Feat(Plate, AddFork): AddFork UI 수정, Plate 리스트 뷰
- AddFork 매장 명 추가
- AddForkSearch 아이템 카테고리 라벨 추가
- 별점 소수점 표기
- Plate뷰 Rx 사용하여 List 표시

+++ FoodFork/UI/AddFork/AddForkStarRate/AddForkStarRateViewModel.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 class AddForkStarRateViewModel {
-    var starRate: Int = 0
+    var starRate: Double = 0
 }
